fixed func declaration

+++ matrices.h
@@ -10,6 +10,6 @@ uint8_t** createMatrix(float prob);
 void printMatrix(int** matrix);
 void freeMatrix(int **matrix);
 int testMatrix(int** matrix_1, int** matrix_2);
-int** multiplyMatrix(int **matrix_1, int **matrix_2);
+int** multiplyMatrix(uint8_t **matrix_1, uint8_t **matrix_2);
 
 #endif //MATRICES_H
